+++ foxes/input/states/__init__.py
@@ -39,5 +39,6 @@ from .newa_states import NEWAStates as NEWAStates
 from .icon_states import ICONStates as ICONStates
 from .ref_point_fields import SectorSimRefPointField as SectorSimRefPointField
 from .meso_micro_field import MesoMicroField as MesoMicroField
+from .binned_states import BinnedStates as BinnedStates
 
 from . import create as create